refactor common steps, independent features
