+++ Jenkinsfile
@@ -1,10 +1,9 @@
 #!groovy
-node('four') {
+node('xenial') {
 	stage('Checkout') {
 	  clean_checkout()
 	}
 	stage('Test') {
-    sh 'gpg2 --version'
 	  sh 'behave -v -x --capture'
 	}
 }
