+++ features/environment.py
@@ -17,9 +17,6 @@ def shell_command(command):
     result.wait()
 
 def before_all(context):
-    # shell_command('gpgconf --kill gpg-agent')
-    # shell_command('gpg-agent --daemon --allow-preset-passphrase --batch --homedir {0}/features/keys'.format(os.getcwd()))
-    # shell_command('/usr/lib/gnupg2/gpg-preset-passphrase --preset --passphrase Iamatestkey 7251332279C83ED15745B9E71880F3E5794267AC')
     pass
 
 def before_feature(context, feature):
@@ -51,10 +48,9 @@ def after_step(context, step):
     pass
 
 def after_scenario(context, scenario):
-    # shell_command('rm -rf {0}'.format(context.mock_developer_dir))
-    # shell_command('rm -rf {0}'.format(context.mock_github_dir))
-    # shell_command('rm -rf {0}'.format(context.gnupghome_dir))
-    pass
+    shell_command('rm -rf {0}'.format(context.mock_developer_dir))
+    shell_command('rm -rf {0}'.format(context.mock_github_dir))
+    shell_command('rm -rf {0}'.format(context.gnupghome_dir))
 
 def after_feature(context, feature):
     pass
